PRSDK-7736: sciclient: Fix for unaligned C66x accesses leading to a hang in sciclient APIS
The C66x does not handle unaligned writes well. This causes the flags settings to overwrite the host ID in the header. This leads to some messages with a valid flag value to hang as the message is never sent to the DMSC.

+++ packages/ti/drv/sciclient/src/sciclient_priv.h
@@ -93,7 +93,7 @@ extern "C" {
  */
 typedef struct
 {
-    uint8_t context;
+    uint32_t context;
     /**< context(sec/non-sec) **/
 
     uint32_t hostId;
